+++ tests/Integration/Model/BothTest.php
@@ -28,9 +28,16 @@ class BothTest
      */
     protected $others;
 
-    public function __construct()
+    /**
+     * @OGM\Relationship(relationshipEntity="BothRel", direction="BOTH", collection=true, type="FRIEND")
+     */
+    protected $friends;
+
+    public function __construct($name)
     {
+        $this->name = $name;
         $this->others = new ArrayCollection();
+        $this->friends = new ArrayCollection();
     }
 
     /**
@@ -60,11 +67,26 @@ class BothTest
         return false;
     }
 
+    public function addOther(BothTest $other)
+    {
+        $this->others->add($other);
+    }
+
     /**
      * @return mixed
      */
     public function getOthers()
     {
         return $this->others;
+    }
+
+    public function addFriend(BothTest $bothTest)
+    {
+        $this->friends->add(new BothRel($this, $bothTest));
+    }
+
+    public function getFriends()
+    {
+        return $this->friends;
     }
 }